refactor(data-summary): integrate functions for region unit counts and laser conditions
- Added functions `extract_region_unit_counts` and `extract_laser_conditions` to the data summary.
- These functions extract additional dataset characteristics to enhance analysis.
- Updated the main script to include new data in the summary generation process.
- Implemented pipeline graph checks for improved logging and traceability.
